+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- Upgraded `typescript` to `6.0.3` for `packages/website-generator` and
+  `website`.
 - Updated Docusaurus to `3.10.2`.
 - Updated `@types/node` to `24.2.0` for `packages/website-generator`.